Refactor exports for chart utils
