+++ packages/react-charts/src/components/ChartBullet/ChartBulletComparativeWarningMeasure.tsx
@@ -2,10 +2,10 @@ import * as React from 'react';
 import hoistNonReactStatics from 'hoist-non-react-statics';
 import { DataGetterPropType, DomainPropType, NumberOrCallback, PaddingProps } from 'victory-core';
 import { VictoryBar } from 'victory-bar';
+import { ChartBulletComparativeMeasure } from './ChartBulletComparativeMeasure';
 import { ChartContainer } from '../ChartContainer';
 import { ChartThemeDefinition } from '../ChartTheme';
-import { getBulletComparativeWarningMeasureTheme } from '../ChartUtils';
-import { ChartBulletComparativeMeasure } from './ChartBulletComparativeMeasure';
+import { getBulletComparativeWarningMeasureTheme } from '../ChartUtils/chart-theme-types';
 
 /**
  * ChartBulletComparativeWarningMeasure renders a dataset as a comparative warning measure.
