+++ packages/react-charts/src/components/ChartBullet/ChartBulletTitle.tsx
@@ -4,7 +4,9 @@ import { ChartContainer } from '../ChartContainer';
 import { ChartLabel } from '../ChartLabel';
 import { ChartLegendPosition } from '../ChartLegend';
 import { ChartBulletStyles, ChartCommonStyles, ChartThemeDefinition } from '../ChartTheme';
-import { getBulletTheme, getBulletLabelX, getBulletLabelY, getPaddingForSide } from '../ChartUtils';
+import { getBulletLabelX, getBulletLabelY } from '../ChartUtils/chart-label';
+import { getPaddingForSide } from '../ChartUtils/chart-padding';
+import { getBulletTheme } from '../ChartUtils/chart-theme-types';
 
 /**
  * ChartBulletTitle renders the bullet chart title.
