+++ packages/react-charts/src/components/ChartBullet/utils/chart-bullet-domain.ts
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { ChartDomain, getDomains } from '../../ChartUtils';
+import { ChartDomain, getDomains } from '../../ChartUtils/chart-domain';
 
 interface ChartBulletDomainInterface {
   comparativeErrorMeasureComponent?: React.ReactElement<any>;
