+++ packages/react-charts/src/components/ChartCursorTooltip/ChartCursorTooltip.tsx
@@ -11,7 +11,8 @@ import {
 import { VictoryTooltip } from 'victory-tooltip';
 import { ChartThemeDefinition } from '../ChartTheme';
 import { ChartTooltip, ChartTooltipProps } from '../ChartTooltip';
-import { getTheme, getCursorTooltipCenterOffset, getCursorTooltipPoniterOrientation } from '../ChartUtils';
+import { getCursorTooltipCenterOffset, getCursorTooltipPoniterOrientation } from '../ChartUtils/chart-tooltip';
+import { getTheme } from '../ChartUtils/chart-theme';
 import { ChartCursorFlyout } from './ChartCursorFlyout';
 
 /**
