+++ packages/react-charts/src/components/ChartLegend/ChartLegend.tsx
@@ -21,7 +21,7 @@ import { ChartContainer } from '../ChartContainer';
 import { ChartLabel } from '../ChartLabel';
 import { ChartPoint } from '../ChartPoint';
 import { ChartThemeDefinition } from '../ChartTheme';
-import { getTheme } from '../ChartUtils';
+import { getTheme } from '../ChartUtils/chart-theme';
 
 export enum ChartLegendOrientation {
   horizontal = 'horizontal',
