+++ packages/react-charts/src/components/ChartTooltip/ChartTooltip.tsx
@@ -11,7 +11,7 @@ import {
 import { VictoryTooltip, VictoryTooltipProps } from 'victory-tooltip';
 import { ChartLabel } from '../ChartLabel';
 import { ChartThemeDefinition } from '../ChartTheme';
-import { getTheme } from '../ChartUtils';
+import { getTheme } from '../ChartUtils/chart-theme';
 
 /**
  * ChartTooltip renders a tooltip component with a set of default events. When ChartTooltip is used as a label component
